+++ src/main/java/com/machineAdmin/models/cg/ModelPerfilYHereda.java
@@ -22,14 +22,14 @@ package com.machineAdmin.models.cg;
  */
 public class ModelPerfilYHereda {
 
-    private Long perfilId;
+    private Integer perfilId;
     private boolean hereda;
 
-    public Long getPerfilId() {
+    public Integer getPerfilId() {
         return perfilId;
     }
 
-    public void setPerfilId(Long profileId) {
+    public void setPerfilId(Integer profileId) {
         this.perfilId = profileId;
     }
